date added in 'Information' item, minor changes in code bcoz of backend-changes

+++ app/src/main/java/com/sahni/rahul/ieee_niec/fragments/InformationFragment.java
@@ -64,7 +64,7 @@ public class InformationFragment extends Fragment implements OnRecyclerViewItemC
 
 //    int mScrollPosition = 0;
 
-    private int mLastItemId;
+    private float mLastItemId;
 
     private CollectionReference mCollectionReference;
